requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
pydantic-settings==2.7.1
pytest==8.3.4
httpx==0.28.1
